Various cleanup

+++ source/contracts/trading/ShareToken.sol
@@ -11,9 +11,8 @@ import 'reporting/IMarket.sol';
 
 contract ShareToken is DelegationTarget, ITyped, Initializable, VariableSupplyToken, IShareToken {
 
-    //FIXME: Delegated contracts cannot currently use string values, so we will need to find a workaround if this hasn't been fixed before we release
     string constant public name = "Shares";
-    uint256 constant public decimals = 0;
+    uint8 constant public decimals = 0;
     string constant public symbol = "SHARE";
 
     IMarket private market;
